Solution for saving referenced objects
Suppose you wish to save a Book object - should the object contain, as
named property, "author_id" referencing its author, which should be
updated whenever a new Author object is passed in to the Book? Makes
more sense instead to grab the $author->id directly when saving, so
Mappers now have a simple structure to do this. So, passing a Book in to
its Mapper::saveEntity() method will cause the Mapper to look for the
Book's Author object, and if it's not a placeholder, take its "id"
property and stick that in its own "author_id" one. Nice.