Switching Form Validation in to... Forms :/
Makes sense now I type that out, but with this "separation of concerns"
it's never 100% clear where things should reside. Are we validating the
data belonging to the Form, or the data going in to a Model, or the data
being passed to a db via a Mapper? Clearly it's all three, but as I've
discovered Forms already implement InputFilter I've shunted the stuff
out of Mapper and in to Form.

Next up: try to get db authentication working.